# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SilentRelay API
  description: |
    Secure end-to-end encrypted messaging API for SilentRelay.
    
    ## Authentication
    Most endpoints require JWT authentication via `Authorization: Bearer <token>` header.
    
    ## Rate Limiting
    - SMS endpoints: 5 requests/minute/IP
    - Auth endpoints: 10 requests/minute/IP  
    - Protected endpoints: 100 requests/minute/user
    
    ## Security
    All communications use end-to-end encryption. The server never has access to message content.
  version: 1.0.0
  contact:
    name: SilentRelay Support
    url: https://silentrelay.com.au
  license:
    name: Proprietary
    url: https://silentrelay.com.au/terms
servers:
  - url: https://silentrelay.com.au/api/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login or registration
...
paths:
...
  /groups/{groupId}/members:
    post:
      tags:
        - Groups
      summary: Add group member
      description: Adds a new member to the group
      operationId: addGroupMember
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
              properties:
                user_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Member added
...